+++ Tasky/View/TaskListView.swift
@@ -22,42 +22,6 @@ struct TaskListView: View {
         .background(Color.black)
     }
 }
-// checkmark.circle.fill
-// circle
-
-struct TaskInfoView: View {
-    let task: Tasky.Task
-    let manager: TaskListViewModel
-    
-    var body: some View {
-        HStack {
-           Image(systemName: imageName())
-                .font(.title2)
-                .imageScale(.large)
-                .foregroundStyle(task.isComplete ? .gray : .purple)
-                .onTapGesture {
-                    manager.changeTaskStatus(task)
-                }
-            VStack(alignment: .leading) {
-                Text(task.title)
-                    .lineLimit(1)
-                    .font(.subheadline)
-                    .foregroundStyle(task.isComplete ? .gray : .white)
-                    .overlay {
-                        if task.isComplete {
-                            Rectangle().frame(height: 1)
-                                .foregroundStyle(.gray)
-                        }
-                    }
-                // TODO: Add Due date here...
-            }
-        }
-    }
-    
-    private func imageName() -> String {
-        task.isComplete ? "checkmark.circle" : "circle"
-    }
-}
 
 
 #Preview {
